Correciones de seguridad y uso de DTOs
Mejora en usuarios por seguridad. Refactor de controladores y servicios para usar DTOs. Unificación a Float en el manejo de precios en toda la aplicacion.

+++ ecommerce/src/main/java/com/tpo_G2/ecommerce/dto/ItemCarritoDTO.java
@@ -8,6 +8,6 @@ import lombok.Data;
 public class ItemCarritoDTO {
     private Long id;
     private int cantidad;
-    private double precioUnitario;
+    private Float precioUnitario;
     private ProductoDTO producto;
 }
